// gym_maxxing/src/app/scheduled_classes/interfaces.ts
export interface ClassInfo {
    classID: number;
    title: string;
    date: string;       
    time: string;       
    trainerName: string;
    classSize: number;
    gymID: number;       
  }
  
  export interface Gym {
    gym_id: number;
    gym_name: string;
    street: string;
    city: string;
    state: string;
    postal_code: number;
    phone_numer: number;
  }
  
  